fix : DeleteDeploymentInputPort findHosting 처리

+++ deployment-service/src/main/java/com/ttalkak/deployment/deployment/application/inputport/DeleteDeploymentInputPort.java
@@ -55,6 +55,9 @@ public class DeleteDeploymentInputPort implements DeleteDeploymentUsecase {
             throw new BusinessException(ErrorCode.KAFKA_PRODUCER_ERROR);
         }
         HostingEntity findHosting = hostingOutputPort.findByProjectIdAndServiceType(deploymentEntity.getProjectId(), deploymentEntity.getServiceType());
+        if(findHosting == null){
+            throw new BusinessException(ErrorCode.NOT_EXISTS_HOSTING);
+        }
         findHosting.delete();
         domainOutputPort.deleteDomainKey(findHosting.getId().toString());
         deploymentEntity.deleteDeployment();
